pyproject.toml
[project]
name = "llm-challenge"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.16.5",
    "fastapi>=0.116.1",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.10.1",
    "uvicorn>=0.35.0",
    "python-dotenv>=1.0.0",
    "asyncpg>=0.30.0",
    "bcrypt>=4.3.0",
    "pyjwt>=2.10.1",
    "streamlit>=1.49.1",
    "azure-storage-blob>=12.26.0",
    "python-multipart>=0.0.20",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "ruff>=0.12.11",
]
